Additions to ship placement screen
-Include a "placed" text/symbol next to each ship that makes it clear
when a ship has been placed or not. Includes adding a "placed"
property to the 'createShip' factory. Prevents re-placing a ship that
has already been placed; this ability will be added in a future
commit
-Add a 'reset board' button that shows on every ship placement screen
-Add 'game-ready' screen once all ships have been placed
-To come: clicking "start game" starts a game; ships can be
re-positioned on the board during ship placement; "reset board" &
"place ships for me" button functionality